feat: add button module, implement input characters to the textarea
